feat: complete tast5 module3 v.1

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        final int FIVE = 5;
+        final double FIVE = 5.0;
         double result = toBeDivided / FIVE;
         System.out.println(result);
     }
